+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - core: pull git repos if already cloned
 - core: permit to change repos folder path
-- docker: add cron support
+- export: add suport for remote export of reports with rclone
+- docker: add cron support with supercronic
+- docker: fix opengrep not found by Python subprocess
 
 ## [0.2] - 2025-10-10
 - core: only clone selected git branch